pyproject.toml
[project]
name = "astro-pipeline"
version = "0.1.0"
description = "Config-driven Kaggle image classification pipeline"
requires-python = ">=3.10"
dependencies = [
  "pyyaml",
  "numpy",
  "pandas",
  "scikit-learn",
  "matplotlib",
  "pillow",
  "tqdm"
]

[tool.setuptools]
packages = ["astro_pipe"]